Doc model

+++ src/model/LineWall.java
@@ -1,12 +1,17 @@
 package src.model;
 
+/**
+ * LineWall est la classe représentant un mur entre deux points.
+ * <br/>
+ * Le premier étant (x1, y1) et le deuxième (x2,y2)
+ */
 public class LineWall
 {
 	private int x1;
 	private int x2;
 	private int y1;
 	private int y2;
-	private int epaisseur;
+	private float epaisseur;
 
 	public LineWall(int x1, int x2, int y1, int y2)
 	{
